Changed how items are sorted and disabled the Thoughts section
Sort order for pages was previously based on filename. Now they are
sorted based on a variable within the pages themselves. This means no
URL changing in the future when I add more content.

The Thoughts section was disabled for now until I have new content to
post. All the content there currently is really old and it makes me
look inactive.